+++ lib/llm_cost_tracker/price_sync/source_result.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module LlmCostTracker
-  module PriceSync
-    SourceResult = Data.define(:prices, :missing_models, :source_version)
-  end
-end